Alterações para rodar as migrations em todos os bancos configurados nas tenants
