+++ src/main/allTenants/educacao-dev.properties
@@ -1,0 +1,5 @@
+name=educacao-dev
+datasource.url=jdbc:postgresql://localhost:5432/educacao-dev
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
