+++ src/main/allTenants/educacao-prod.properties
@@ -1,0 +1,5 @@
+name=educacao-prod
+datasource.url=jdbc:postgresql://localhost:5432/educacao-prod
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
