+++ src/main/allTenants/educacao-teste.properties
@@ -1,0 +1,5 @@
+name=educacao-teste
+datasource.url=jdbc:postgresql://localhost:5432/educacao-teste
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
