+++ src/main/allTenants/educacenso-dev.properties
@@ -1,0 +1,5 @@
+name=educacenso-dev
+datasource.url=jdbc:postgresql://localhost:5432/educacenso-dev
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
