+++ src/main/allTenants/educacenso-prod.properties
@@ -1,0 +1,5 @@
+name=educacenso-prod
+datasource.url=jdbc:postgresql://localhost:5432/educacenso-prod
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
