+++ src/main/allTenants/educacenso-teste.properties
@@ -1,0 +1,5 @@
+name=educacenso-teste
+datasource.url=jdbc:postgresql://localhost:5432/educacenso-teste
+datasource.username=postgres
+datasource.password=postgres
+datasource.driver-class-name=org.postgresql.Driver
