--- src/main/allTenants/tenant_1.properties
@@ -1,5 +1,0 @@
-name=tenant_1
-datasource.url=jdbc:postgresql://localhost:5432/educacao-dev
-datasource.username=postgres
-datasource.password=postgres
-datasource.driver-class-name=org.postgresql.Driver
