--- src/main/allTenants/tenant_2.properties
@@ -1,5 +1,0 @@
-name=tenant_2
-datasource.url=jdbc:postgresql://localhost:5432/educacao
-datasource.username=postgres
-datasource.password=postgres
-datasource.driver-class-name=org.postgresql.Driver
