--- src/main/allTenants/tenant_3.properties
@@ -1,5 +1,0 @@
-name=tenant_3
-datasource.url=jdbc:postgresql://localhost:5432/educacenso
-datasource.username=postgres
-datasource.password=postgres
-datasource.driver-class-name=org.postgresql.Driver
