+++ src/main/java/br/com/coreeduc/adapters/configuration/BeanConfiguration.java
@@ -11,6 +11,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 
 @Configuration
 @ComponentScan(basePackageClasses = CoreEducApplication.class)
@@ -37,6 +38,11 @@ public class BeanConfiguration {
     @Bean
     public PageRequest pageRequest() {
         return PageRequest.of(INICIO_PAGES, default_quantidade_pages);
+    }
+
+    @Bean
+    InterceptorRegistry InterceptorRegistry( ) {
+        return new InterceptorRegistry();
     }
 
 }
