+++ src/main/java/br/com/coreeduc/adapters/configuration/WebConfiguration.java
@@ -1,0 +1,20 @@
+package br.com.coreeduc.adapters.configuration;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+@Getter
+public class WebConfiguration implements WebMvcConfigurer {
+
+    private final TenantInterceptor tenantInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addWebRequestInterceptor(tenantInterceptor);
+    }
+}
