+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/AuthorizationException.java
@@ -1,0 +1,15 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+public class AuthorizationException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public AuthorizationException(String msg) {
+        super(msg);
+    }
+
+    public AuthorizationException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}
