+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/DataIntegrityException.java
@@ -1,0 +1,15 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+public class DataIntegrityException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public DataIntegrityException(String msg) {
+        super(msg);
+    }
+
+    public DataIntegrityException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}
