+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/FileException.java
@@ -1,0 +1,15 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+public class FileException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public FileException(String msg) {
+        super(msg);
+    }
+
+    public FileException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}
