+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/ObjectNotFoundException.java
@@ -1,0 +1,15 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+public class ObjectNotFoundException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public ObjectNotFoundException(String msg) {
+        super(msg);
+    }
+
+    public ObjectNotFoundException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}
