+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/ValidationError.java
@@ -1,0 +1,22 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ValidationError extends StandardError {
+    private static final long serialVersionUID = 1L;
+
+    private List<FieldMessage> errors = new ArrayList<>();
+
+    public ValidationError(Long timestamp, Integer status, String error, String message, String path) {
+        super(timestamp, status, error, message, path);
+    }
+
+    public List<FieldMessage> getErrors() {
+        return errors;
+    }
+
+    public void addError(String fieldName, String messagem) {
+        errors.add(new FieldMessage(fieldName, messagem));
+    }
+}
