+++ src/main/java/br/com/coreeduc/adapters/inbound/exceptions/ValidationException.java
@@ -1,0 +1,15 @@
+package br.com.coreeduc.adapters.inbound.exceptions;
+
+public class ValidationException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public ValidationException(String msg) {
+        super(msg);
+    }
+
+    public ValidationException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}
