--- src/main/java/br/com/coreeduc/adapters/inbound/resources/exceptions/DataIntegrityException.java
@@ -1,15 +1,0 @@
-package br.com.coreeduc.adapters.inbound.resources.exceptions;
-
-public class DataIntegrityException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    public DataIntegrityException(String msg) {
-        super(msg);
-    }
-
-    public DataIntegrityException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-}
