--- src/main/java/br/com/coreeduc/adapters/inbound/resources/exceptions/ValidationError.java
@@ -1,22 +1,0 @@
-package br.com.coreeduc.adapters.inbound.resources.exceptions;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ValidationError extends StandardError {
-    private static final long serialVersionUID = 1L;
-
-    private List<FieldMessage> errors = new ArrayList<>();
-
-    public ValidationError(Long timestamp, Integer status, String error, String message, String path) {
-        super(timestamp, status, error, message, path);
-    }
-
-    public List<FieldMessage> getErrors() {
-        return errors;
-    }
-
-    public void addError(String fieldName, String messagem) {
-        errors.add(new FieldMessage(fieldName, messagem));
-    }
-}
