--- src/main/java/br/com/coreeduc/adapters/inbound/resources/exceptions/ValidationException.java
@@ -1,15 +1,0 @@
-package br.com.coreeduc.adapters.inbound.resources.exceptions;
-
-public class ValidationException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    public ValidationException(String msg) {
-        super(msg);
-    }
-
-    public ValidationException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-}
