+++ src/main/resources/application-dev.properties
@@ -3,9 +3,15 @@ server.port=8085
 eureka.client.serviceUrl.defaultZone=http://localhost:8081/eureka
 eureka.instance.prefer-ip-address=true
 eureka.instance.hostname=localhost
-defaultTenant=tenant_1
-spring.rabbitmq.addresses=amqps://jwazxyfh:R6H1_jEFbcIFV_d3TmYrTfPIJyWs6697@beaver.rmq.cloudamqp.com/jwazxyfh
+
+tenants.datasources.educacao-dev=educacao-dev
+tenants.datasources.educacenso-dev=educacenso-dev
+defaultTenant=educacao-dev
+
+spring.rabbitmq.addresses=amqps://nwxtjpjc:q_qeE-qLL9GVCLjfpYSlaF0jHhKTKNYX@beaver.rmq.cloudamqp.com/nwxtjpjc
 spring.rabbitmq.queue=ms.educacenso
 rabbitmq.exchange.name=educacenso
 rabbitmq.routing.key=educacenso
-
+spring.flyway.enabled=false
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults = false
