+++ src/main/resources/application-prod.properties
@@ -1,7 +1,16 @@
 host-swagger=localhost
 server.port=8085
-spring.datasource.url=jdbc:postgresql://localhost:5432/educacao
-spring.datasource.username=postgres
-spring.datasource.password=postgres
 eureka.client.serviceUrl.defaultZone=http://localhost:8081/eureka
-defaultTenant=tenant_2
+eureka.instance.prefer-ip-address=true
+eureka.instance.hostname=localhost
+
+tenants.datasources.educacao-prod=educacao-prod
+tenants.datasources.educacenso-prod=educacenso-prod
+defaultTenant=educacao-prod
+
+spring.rabbitmq.addresses=amqps://nwxtjpjc:q_qeE-qLL9GVCLjfpYSlaF0jHhKTKNYX@beaver.rmq.cloudamqp.com/nwxtjpjc
+spring.rabbitmq.queue=ms.educacenso
+rabbitmq.exchange.name=educacenso
+rabbitmq.routing.key=educacenso
+spring.flyway.enabled=false
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
