+++ src/main/resources/application-test.properties
@@ -1,0 +1,16 @@
+host-swagger=localhost
+server.port=8085
+eureka.client.serviceUrl.defaultZone=http://localhost:8081/eureka
+eureka.instance.prefer-ip-address=true
+eureka.instance.hostname=localhost
+
+tenants.datasources.educacao-teste=educacao-teste
+tenants.datasources.educacenso-teste=educacenso-teste
+defaultTenant=educacao-teste
+
+spring.rabbitmq.addresses=amqps://nwxtjpjc:q_qeE-qLL9GVCLjfpYSlaF0jHhKTKNYX@beaver.rmq.cloudamqp.com/nwxtjpjc
+spring.rabbitmq.queue=ms.educacenso
+rabbitmq.exchange.name=educacenso
+rabbitmq.routing.key=educacenso
+spring.flyway.enabled=false
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
