+++ src/main/resources/application.properties
@@ -24,3 +24,4 @@ spring.jpa.database=default
 
 
 
+
