+++ src/main/resources/db/migration/V77__add_coluna_tenant_user_.sql
@@ -1,2 +1,2 @@
 ALTER TABLE tb_user ADD tenant varchar NULL;
-update tb_user set tenant = 'tenant_1';
+update tb_user set tenant = 'educacao-dev';
