+++ src/main/resources/db/migration/V78__create_new_user.sql
@@ -12,4 +12,4 @@ insert into TB_USER (ds_login,
 
 values ('SECRE', 'Secretario', '93140441061',
         '$2a$10$3XhhrCOssmvwSty2jH2CveJFyaybagNXzND2WC0SPFGcROvZt5ze.', '145875647', 'secretario@gmail.com', '2001-03-16',
-        '62999077331', now(), 1, 1, 'tenant_2');
+        '62999077331', now(), 1, 1, 'educacao-prod');
