+++ src/test/java/br/com/coreeduc/adapters/inbound/resources/exceptions/AuthorizationExceptionTest.java
@@ -1,5 +1,6 @@
 package br.com.coreeduc.adapters.inbound.resources.exceptions;
 
+import br.com.coreeduc.adapters.inbound.exceptions.AuthorizationException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
